Have a dataflowops wrapper around the libattentionhelper function that queries the required workspace size

+++ libdataflowops/include/dataflow_ops.h
@@ -83,7 +83,7 @@ int dataflow_submit_matmul(Dataflow_Handle * handle, int stream_id,
 // returns the required workspace size for the attention op
 // exported by libattention_helper.so
 
-int dataflow_get_attention_workspace_size(Dataflow_Handle * handle, DataflowDatatype dtype, int is_training, 
+uint64_t dataflow_get_attention_workspace_size(Dataflow_Handle * handle, DataflowDatatype dtype, int is_training, 
 											int num_q_heads, int num_kv_heads, int head_dim, 
 											int max_chunk_size, int max_seq_len, int max_seqs_in_chunk,
 											int is_causal);
